Sorted out broken debug gui text things and added a creature counter.

+++ Assets/Scripts/Environment/Ether.cs
@@ -72,7 +72,7 @@ public class Ether : MonoBehaviour {
 		return foodbitEnergy;
 	}
 	
-	public double getEnergy() {
+	public int getEnergy() {
 		return energy;
 	}
